Rename `events` -> `event`
